refactor(ledger): collocate the metadata-index route registrars :hammer:
Both metadata-index registrars move out of metadata_handler.go into
metadata_routes.go, leaving the handler file with the handler struct, the
request/response envelopes and the cores.

No route, authz tuple, operation ID or path changed.
